Revert "Replace "= size_t(-1)" with " -1" for size_t"
This reverts commit 8ecfe21eb0232a295e4d36a7e885f7201b080d55.

+++ include/OOBase/BTree.h
@@ -717,7 +717,7 @@ namespace OOBase
 					else
 						end = mid;
 				}
-				return -1;
+				return size_t(-1);
 			}
 
 			bool insert_i(BTreeImpl<K,V,Compare,B,Allocator>* tree, const Pair<K,V>& value, size_t parent_pos)
