+++ include/OOBase/String.h
@@ -80,7 +80,7 @@ namespace OOBase
 	class ScopedStringImpl : public NonCopyable
 	{
 	public:
-		static const size_t npos = -1;
+		static const size_t npos = size_t(-1);
 
 		ScopedStringImpl() : m_data(), m_len(0)
 		{}
